Rename package in order to be compatible with the other modules
