+++ xchange-coinmate/src/main/java/info/bitrich/xchangestream/coinmate/CoinmateStreamingAccountService.java
@@ -1,7 +1,7 @@
-package info.bitrich.xchange.coinmate;
+package info.bitrich.xchangestream.coinmate;
 
 import com.fasterxml.jackson.core.type.TypeReference;
-import info.bitrich.xchange.coinmate.dto.CoinmateWebsocketBalance;
+import info.bitrich.xchangestream.coinmate.dto.CoinmateWebsocketBalance;
 import info.bitrich.xchangestream.core.StreamingAccountService;
 import info.bitrich.xchangestream.service.netty.StreamingObjectMapperHelper;
 import info.bitrich.xchangestream.service.pusher.PusherStreamingService;
@@ -57,7 +57,7 @@ public class CoinmateStreamingAccountService implements StreamingAccountService 
         });
     }
 
-    private Observable<Map<String,CoinmateWebsocketBalance>> getCoinmateBalances(){
+    private Observable<Map<String, CoinmateWebsocketBalance>> getCoinmateBalances(){
         String channelName = "private-user_balances-" + userId;
 
         return service.subscribeChannel(channelName,"user_balances")
